requirements.txt
# Core dependencies
boto3>=1.26.0
python-dotenv>=1.0.0
PyYAML>=6.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
wandb>=0.15.0
click>=8.0.0

# Kaggle API
kaggle>=1.5.0

# Deep Learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.0

# Image augmentation
albumentations>=1.3.0

# Scientific computing
scipy>=1.10.0
